backend/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
python-dotenv
langchain
langchain-groq
langchain-core
pydantic
requests